add sep files
